refactor: relocate medication shared components
