+++ src/views/content/medications/NsSideeffect.vue
