+++ src/views/content/medications/glucagon/ContentGlucagon.vue
@@ -100,14 +100,14 @@ import NsContentGroup from '@/components/NsContentGroup.vue'
 import NsQuicktip from '@/components/NsQuicktip.vue'
 import NsList from '@/components/NsList.vue'
 import NsListItem from '@/components/NsListItem.vue'
-import NsContraindication from '@/components/NsContraindication.vue'
-import NsSideeffect from '@/components/NsSideeffect.vue'
-import NsPackage from '@/components/NsPackage.vue'
-import NsDosageIndication from '@/components/NsDosageIndication.vue'
-import NsDosageUsage from '@/components/NsDosageUsage.vue'
-import NsDosage from '@/components/NsDosage.vue'
-import NsPharmacokinetics from '@/components/NsPharmacokinetics.vue'
-import NsPharmacodynamics from '@/components/NsPharmacodynamics.vue'
+import NsContraindication from '@/views/content/medications/NsContraindication.vue'
+import NsSideeffect from '@/views/content/medications/NsSideeffect.vue'
+import NsPackage from '@/views/content/medications/NsPackage.vue'
+import NsDosageIndication from '@/views/content/medications/NsDosageIndication.vue'
+import NsDosageUsage from '@/views/content/medications/NsDosageUsage.vue'
+import NsDosage from '@/views/content/medications/NsDosage.vue'
+import NsPharmacokinetics from '@/views/content/medications/NsPharmacokinetics.vue'
+import NsPharmacodynamics from '@/views/content/medications/NsPharmacodynamics.vue'
 import TextMono from '@/components/TextMono.vue'
 import TextUnderline from '@/components/TextUnderline.vue'
